Forcing the buildpack to use node 14
The heroku-nodejs-plugin no longer supports node 14 in their latest
release this caused our production builds to fail because of a 404
error when attempting to retrieve the node 14 tarball. Heroku support
shared this is not an offical buildpack and can't assist. I've forced
the download logic to reference an older release and use node 14.